Move the test to a functional test
